time partition of 1/20 with step of 1 ----> partition of 1 with step of 1/20, in rk changed to dtype to float32 to save some ram, and chaning the number of dimensions of A1,A2,F to match the intuitive and not the programatically practical. A in R^(WxHx1) -----> A in R^(WxH)

+++ src/rk.py
@@ -4,7 +4,7 @@ import math
 
 def solve(_dYdt,Y0,time_range,args=(),steps=1):
     #main
-    Y=np.zeros([len(time_range),len(Y0)])
+    Y=np.zeros([len(time_range),len(Y0)],dtype=np.float32)
     Y[0]=Y0#<---initial conditions
     def dYdt(Y,t,_args):
     	Y[Y<0]=0#this is to make rk work
